Move the users seeder to internal
